+++ src/core/tamper/backslashes.py
@@ -15,35 +15,31 @@ For more see the file 'readme/COPYING' for copying permission.
 
 import re
 import sys
-from src.utils import menu
 from src.utils import settings
 
 """
-About: Adds back slashes ("\") between the characters of the generated payloads.
+About: Adds back slashes (\) between the characters in a given payload.
 Notes: This tamper script works against Unix-like target(s).
 """
 
 __tamper__ = "backslashes"
 
+global obf_char
+
 if not settings.TAMPER_SCRIPTS[__tamper__]:
+  obf_char = "\\"
   settings.TAMPER_SCRIPTS[__tamper__] = True
 
 def tamper(payload):
   def add_back_slashes(payload):
-    settings.TAMPER_SCRIPTS[__tamper__] = True
-    obf_char = "\\"
-    payload = re.sub(r'([b-zD-Z])', lambda x: obf_char + x[0], payload)
+    payload = re.sub(settings.TAMPER_MODIFICATION_LETTERS, lambda x: obf_char + x[0], payload)
     for word in settings.IGNORE_TAMPER_TRANSFORMATION:
       _ = obf_char.join(word[i:i+1] for i in range(-1, len(word), 1))
       if _ in payload:
         payload = payload.replace(_,_.replace(obf_char, ""))
     return payload
-
   if settings.TARGET_OS != settings.OS.WINDOWS:
-    if settings.EVAL_BASED_STATE != False:
-      return payload
-    else:
-      return add_back_slashes(payload)
+    return add_back_slashes(payload)
   else:
     return payload