+++ src/core/tamper/backticks.py
@@ -18,7 +18,7 @@ from src.utils import settings
 
 
 """
-About: Uses backticks instead of "$()" for commands substitution on the generated payloads.
+About: Uses backticks (`) instead of "$()" for commands substitution in a given payload.
 Notes: This tamper script works against Unix-like target(s).
 """
 
@@ -28,7 +28,6 @@ if not settings.TAMPER_SCRIPTS[__tamper__]:
   settings.TAMPER_SCRIPTS[__tamper__] = True
 
 def tamper(payload):
-  settings.TAMPER_SCRIPTS[__tamper__] = True
   if not menu.options.alter_shell and not settings.TARGET_OS == settings.OS.WINDOWS:
     settings.USE_BACKTICKS = True
     settings.CMD_SUB_PREFIX = settings.CMD_SUB_SUFFIX = "`"