+++ src/core/tamper/base64encode.py
@@ -15,8 +15,8 @@ For more see the file 'readme/COPYING' for copying permission.
 
 import sys
 import base64
-from src.thirdparty.six.moves import urllib as _urllib
 from src.utils import settings
+from src.thirdparty.six.moves import urllib as _urllib
 
 """
 About: Base64 all characters in a given payload.
@@ -31,11 +31,8 @@ if not settings.TAMPER_SCRIPTS[__tamper__]:
 def tamper(payload):
   if len(settings.WHITESPACES) != 0 and settings.WHITESPACES[0] == _urllib.parse.quote_plus(settings.SINGLE_WHITESPACE):
     err_msg = "Tamper script '" +  __tamper__  + "' is unlikely to work combined with the tamper script 'space2plus'."
-    if settings.VERBOSITY_LEVEL == 0:
-      settings.print_data_to_stdout(settings.SINGLE_WHITESPACE)
     settings.print_data_to_stdout(settings.print_critical_msg(err_msg))
     raise SystemExit()
-
   else:
     payload = _urllib.parse.unquote(payload)
     payload = base64.b64encode(payload.encode())