+++ src/core/tamper/caret.py
@@ -19,7 +19,7 @@ from src.utils import menu
 from src.utils import settings
 
 """
-About: Adds caret symbol (^) between the characters of the generated payloads.
+About: Adds caret symbol (^) between the characters in a given payload.
 Notes: This tamper script works against windows targets.
 """
 
@@ -30,7 +30,6 @@ if not settings.TAMPER_SCRIPTS[__tamper__]:
 
 def tamper(payload):
   def add_caret_symbol(payload):
-    settings.TAMPER_SCRIPTS[__tamper__] = True
     if re.compile(r"\w+").findall(payload):
       long_string = ""
       if len(max(re.compile(r"\w+").findall(payload), key=lambda word: len(word))) >= 5000:
@@ -39,19 +38,15 @@ def tamper(payload):
               "^^": "^",
               '"^t""^o""^k""^e""^n""^s"': '"t"^"o"^"k"^"e"^"n"^"s"',
               '^t^o^k^e^n^s': '"t"^"o"^"k"^"e"^"n"^"s"',
-              re.sub(r'([b-zD-Z])', r'^\1', long_string) : long_string.replace("^", "")
+              re.sub(settings.TAMPER_MODIFICATION_LETTERS, r'^\1', long_string) : long_string.replace("^", "")
             }
-      payload = re.sub(r'([b-zD-Z])', r'^\1', payload)
+      payload = re.sub(settings.TAMPER_MODIFICATION_LETTERS, r'^\1', payload)
       rep = dict((re.escape(k), v) for k, v in rep.items())
       pattern = re.compile("|".join(rep.keys()))
       payload = pattern.sub(lambda m: rep[re.escape(m.group(0))], payload)
     return payload
-
   if settings.TARGET_OS == settings.OS.WINDOWS:
-    if settings.EVAL_BASED_STATE != False:
-      return payload
-    else:
-      return add_caret_symbol(payload)
+    return add_caret_symbol(payload)
   else:
     return payload