+++ src/core/tamper/dollaratsigns.py
@@ -17,20 +17,21 @@ import re
 from src.utils import settings
 
 """
-About: Adds dollar sign followed by an at-sign ($@) between the characters of the generated payloads.
+About: Adds dollar sign followed by an at-sign ($@) between the characters in a given payload.
 Notes: This tamper script works against Unix-like target(s).
 """
 
 __tamper__ = "dollaratsigns"
 
+global obf_char
+
 if not settings.TAMPER_SCRIPTS[__tamper__]:
+  obf_char = "$@"
   settings.TAMPER_SCRIPTS[__tamper__] = True
 
 def tamper(payload):
   def add_dollar_at_signs(payload):
-    settings.TAMPER_SCRIPTS[__tamper__] = True
-    obf_char = "$@"
-    payload = re.sub(r'([e-zE-Z])', lambda x: obf_char + x[0], payload)
+    payload = re.sub(settings.TAMPER_MODIFICATION_LETTERS, lambda x: obf_char + x[0], payload)
     for word in settings.IGNORE_TAMPER_TRANSFORMATION:
       _ = obf_char.join(word[i:i+1] for i in range(-1, len(word), 1))
       if _ in payload: