+++ src/core/tamper/hexencode.py
@@ -31,11 +31,8 @@ if not settings.TAMPER_SCRIPTS[__tamper__]:
 def tamper(payload):
   if len(settings.WHITESPACES) != 0 and settings.WHITESPACES[0] == _urllib.parse.quote_plus(settings.SINGLE_WHITESPACE):
     err_msg = "Tamper script '" +  __tamper__  + "' is unlikely to work combined with the tamper script 'space2plus'."
-    if settings.VERBOSITY_LEVEL == 0:
-      settings.print_data_to_stdout(settings.SINGLE_WHITESPACE)
     settings.print_data_to_stdout(settings.print_critical_msg(err_msg))
     raise SystemExit()
-
   else:
     payload = _urllib.parse.unquote(payload)
     encoded_payload, _ = hexencode(payload)