+++ src/core/tamper/multiplespaces.py
@@ -17,16 +17,19 @@ import random
 from src.utils import settings
 
 """
-About: Adds multiple spaces around OS commands
+About: Adds multiple spaces around operating system commands in a given payload.
 Notes: Useful to bypass very weak and bespoke web application firewalls that has poorly written permissive regular expressions.
 """
 
 __tamper__ = "multiplespaces"
 
+if not settings.TAMPER_SCRIPTS[__tamper__]:
+  settings.TAMPER_SCRIPTS[__tamper__] = True
+
 def tamper(payload):
-  if not (settings.TAMPER_SCRIPTS[__tamper__]):
-    settings.TAMPER_SCRIPTS[__tamper__] = True
+  if settings.TAMPER_SCRIPTS[__tamper__]:
     for i in range(0, len(settings.WHITESPACES)):
       settings.WHITESPACES[i] = settings.WHITESPACES[i] * random.randrange(3, 8)
   return payload
+
 # eof