+++ src/core/tamper/space2ifs.py
@@ -17,7 +17,7 @@ from src.utils import settings
 from src.thirdparty.six.moves import urllib as _urllib
 
 """
-About: Replaces space character ('%20') with the internal field separator ('$IFS').
+About: Replaces space character (%20) with the internal field separator ($IFS) in a given payload.
 The internal field separator refers to a variable which defines the character
 or characters used to separate a pattern into tokens for some operations.
 Notes: This tamper script works against Unix-like target(s).
@@ -33,8 +33,7 @@ def tamper(payload):
   if len(settings.WHITESPACES) != 0:
     if space2ifs in settings.WHITESPACES[0] and settings.EVAL_BASED_STATE != False:
       settings.WHITESPACES[0] = space2ifs
-    if settings.TARGET_OS != settings.OS.WINDOWS:
-      settings.TAMPER_SCRIPTS[__tamper__] = True
+    if settings.TARGET_OS != settings.OS.WINDOWS: 
       if settings.WHITESPACES[0] == _urllib.parse.quote(settings.SINGLE_WHITESPACE):
         settings.WHITESPACES[0] = space2ifs
       elif space2ifs not in settings.WHITESPACES: